add check sensor state to ESP8266 sender and Roller boards

+++ pythonScript/main_controller.py
@@ -11,6 +11,7 @@ state_counter = 0
 
 del_cfile_flag = 0
 del_rfile_flag = 0
+del_sfile_flag = 0
 send_file = 0
 main_state = 0
 byte_silo_number = b'9'
@@ -69,8 +70,16 @@ while True:
                             if byte_silo_number == b'c':
                                 del_cfile_flag = 1
                                 time.sleep(0.5)
+                            if byte_silo_number == b's':
+                                del_sfile_flag = 1
+                                time.sleep(0.5)
                 if del_cfile_flag == 1:
                     if not os.path.isfile(path_error_file):
+                        if os.path.isfile(silo_file):
+                            os.remove(silo_file)
+                            print('remove c file')
+                        main_state = 1
+                        del_cfile_flag = 0
                         # reset conveyor status
                         ser.write(b'z\n')
                         time.sleep(1)
@@ -101,6 +110,54 @@ while True:
                         if conveyor_status == b'5':
                             print("Conveyor run complete")
                             del_cfile_flag = 0
+                            main_state = 1
+                            if os.path.isfile(silo_file):
+                                os.remove(silo_file)
+                                print('remove c file')
+                        if conveyor_status == b'6':
+                            print("Conveyor in Empty state")
+                            with open(path_error_file, 'a+') as error_file:
+                                error_file.write("Empty")
+                        if conveyor_status == b'7':
+                            pass
+
+                if del_sfile_flag == 1:
+                    if not os.path.isfile(path_error_file):
+                        if os.path.isfile(silo_file):
+                            os.remove(silo_file)
+                            print('remove s file')
+                        main_state = 1
+                        del_sfile_flag = 0
+                        # reset conveyor status
+                        ser.write(b'z\n')
+                        time.sleep(1)
+                        ser.write(b'y\n')
+                        time.sleep(3)
+                        ser.write(b'u\n')
+                        time.sleep(3)
+                        # check conveyor status
+                        ser.write(b'x\n')
+                        time.sleep(0.5)
+                        conveyor_status = ser.readline()
+                        conveyor_status = conveyor_status.strip()
+                        print(conveyor_status)
+                        if conveyor_status == b'0':
+                            print("Conveyor in idle state")
+                        if conveyor_status == b'1':
+                            print("Conveyor in waiting state")
+                        if conveyor_status == b'2':
+                            print("Conveyor in running state")
+                        if conveyor_status == b'3':
+                            print("Conveyor in full state")
+                            with open(path_error_file, 'a+') as error_file:
+                                error_file.write("Full")
+                        if conveyor_status == b'4':
+                            print("Conveyor in stuck state")
+                            with open(path_error_file, 'a+') as error_file:
+                                error_file.write("Stuck")
+                        if conveyor_status == b'5':
+                            print("Conveyor run complete")
+                            del_sfile_flag = 0
                             main_state = 1
                             if os.path.isfile(silo_file):
                                 os.remove(silo_file)
